validate api

+++ app/Http/Controllers/CountryController.php
@@ -5,6 +5,8 @@ namespace App\Http\Controllers;
 use App\Models\Country;
 use Illuminate\Http\Request;
 
+use Illuminate\Support\Facades\Validator;
+
 class CountryController extends Controller
 {
     /**
@@ -27,28 +29,48 @@ class CountryController extends Controller
         if(is_null($country) ) {
             return response()->json(['error' => true, 'message' => 'Not found'], 404);
         }
-        return response()->json(Country::find($id), 200);
+        return response()->json($country, 200);
     }
 
     public function countrySave(Request $request) {
+        $rules = [
+            'iso' => 'required|min:2|max:2',
+            'name' => 'required|min:3',
+            'name_en' => 'required|min:3',
+        ];
+        $validator = Validator::make($request->all(), $rules);
+        if($validator->fails()) {
+            return response()->json($validator->errors(), 400);
+        }
         $country = Country::create($request->all());
         return response()->json($country, 201);
     }
 
-    /**
-     * Show the form for editing the specified resource.
-     *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
-    public function countryEdit(Request $request, Country $country)
+    public function countryEdit(Request $request, $id)
     {
+        $rules = [
+            'alias' => 'min:2|max:2',
+            'name' => 'required|min:3',
+            'name_en' => 'required|min:3',
+        ];
+        $validator = Validator::make($request->all(), $rules);
+        if($validator->fails()) {
+            return response()->json($validator->errors(), 400);
+        }
+        $country = Country::find($id);
+        if(is_null($country) ) {
+            return response()->json(['error' => true, 'message' => 'Not found'], 404);
+        }
         $country->update($request->all());
-        return response()->json($country, 200);
+        return response()->json($id, 200);
     }
 
-    public function countryDelete(Request $request, Country $country)
+    public function countryDelete(Request $request, $id)
     {
+        $country = Country::find($id);
+        if(is_null($country) ) {
+            return response()->json(['error' => true, 'message' => 'Not found'], 404);
+        }
         $country->delete();
         return response()->json('', 204);
     }
